refactor: add namespace nesting level, move to different folder
